pyproject.toml
[project]
name = "mbf-singlecell-plotter"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.14"
dependencies = [
 "anndata>=0.12.6",
 "natsort>=8.4.0",
 "numpy>=2.4.3",
 "pandas>=3.0.1",
 "pillow>=12.1.1",
 "plotnine>=0.15.3",
]

# [project.scripts]
# hello = "hello_world:hello"

[build-system]
requires = ["uv_build>=0.10.0,<0.11.0"]
build-backend = "uv_build"

[dependency-groups]
dev = [
    "pytest>=9.0.2",
    "scikit-image>=0.24",
    "scipy>=1.14",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src", "tests"]
